+++ README.rst
@@ -1,28 +1,25 @@
-Phantom
+XDesign
 #######
 
-.. image:: https://readthedocs.org/projects/phantom/badge/?version=latest
-   :target: http://phantom.readthedocs.io/en/latest/?badge=latest
+.. image:: https://readthedocs.org/projects/xdesign/badge/?version=latest
+   :target: http://xdesign.readthedocs.io/en/latest/?badge=latest
    :alt: Read the Docs
 
-.. image:: https://travis-ci.org/tomography/phantom.svg?branch=master
-   :target: https://travis-ci.org/tomography/phantom
+.. image:: https://travis-ci.org/tomography/xdesign.svg?branch=master
+   :target: https://travis-ci.org/tomography/xdesign
    :alt: Travis CI
 
-.. image:: https://coveralls.io/repos/github/tomography/phantom/badge.svg?branch=master 
-   :target: https://coveralls.io/github/tomography/phantom?branch=master
+.. image:: https://coveralls.io/repos/github/tomography/xdesign/badge.svg?branch=master 
+   :target: https://coveralls.io/github/tomography/xdesign?branch=master
    :alt: Coveralls
    
-.. image:: https://codeclimate.com/github/tomography/phantom/badges/gpa.svg
-   :target: https://codeclimate.com/github/tomography/phantom
+.. image:: https://codeclimate.com/github/tomography/xdesign/badges/gpa.svg
+   :target: https://codeclimate.com/github/tomography/xdesign
    :alt: Code Climate
 
-**Phantom** is an open-source Python package for generating
-simulation phantoms for evaluating tomographic data processing
-and image reconstruction.
-
-.. DANGER::
-   Package under construction!
+**XDesign** is an open-source Python package for generating
+configurable simulation phantoms for benchmarking tomographic 
+image reconstruction.
 
 Features
 ========
@@ -34,13 +31,13 @@ Features
 Contribute
 ==========
 
-* Issue Tracker: https://github.com/tomography/phantom/issues
-* Documentation: https://github.com/tomography/phantom/tree/master/doc
-* Source Code: https://github.com/tomography/phantom/tree/master/phantom
-* Tests: https://github.com/tomography/phantom/tree/master/test
+* Issue Tracker: https://github.com/tomography/xdesign/issues
+* Documentation: https://github.com/tomography/xdesign/tree/master/doc
+* Source Code: https://github.com/tomography/xdesign/tree/master/xdesign
+* Tests: https://github.com/tomography/xdesign/tree/master/test
 
 License
 =======
 
 The project is licensed under the 
-`BSD-3 <https://github.com/tomography/phantom/blob/master/LICENSE.txt>`_ license.
+`BSD-3 <https://github.com/tomography/xdesign/blob/master/LICENSE.txt>`_ license.
