+++ docs/source/index.rst
@@ -1,12 +1,9 @@
-Phantom
+XDesign
 #######
 
-**Phantom** is an open-source Python package for generating
+**XDesign** is an open-source Python package for generating
 configurable simulation phantoms for benchmarking tomographic 
 image reconstruction.
-
-.. DANGER::
-   Package under construction!
 
 .. toctree::
    :maxdepth: 1
@@ -24,16 +21,16 @@ Features
 Contribute
 ==========
 
-* Issue Tracker: https://github.com/tomography/phantom/issues
-* Documentation: https://github.com/tomography/phantom/tree/master/doc
-* Source Code: https://github.com/tomography/phantom
-* Tests: https://github.com/tomography/phantom/tree/master/test
+* Issue Tracker: https://github.com/tomography/xdesign/issues
+* Documentation: https://github.com/tomography/xdesign/tree/master/doc
+* Source Code: https://github.com/tomography/xdesign/tree/master/xdesign
+* Tests: https://github.com/tomography/xdesign/tree/master/test
 
 License
 =======
 
 The project is licensed under the 
-`BSD-3 <https://github.com/tomography/phantom/blob/master/LICENSE.txt>`_ license.
+`BSD-3 <https://github.com/tomography/xdesign/blob/master/LICENSE.txt>`_ license.
 
 Indices and tables
 ==================
