--- phantom/constants.py
@@ -1,43 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Constants in cgs units.
-
-AVOGADRO_NUMBER : float
-    Avagadro constant [1/mol]
-BOLTZMANN_CONSTANT : float
-    Boltzmann constant [erg/k]
-CLASSICAL_ELECTRON_RADIUS : float
-    Classical electron radius [cm]
-ELECTRONIC_CHARGE : float
-    Electronic charge [esu]
-ELECTRON_VOLT : float
-    Electron volt (keV) [erg]
-ELECTRON_MASS : float
-    Electron mass [g]
-FINE_STRUCTURE_CONSTANT : float
-    Fine structure constant
-PLANCK_CONSTANT : float
-    Reduced planck's constant [keV*s]
-PROTON_MASS : float
-    Proton mass [g]
-SPEED_OF_LIGHT : float
-    Speed of light in vacuum [cm/s]
-THOMPSON_CROSS_SECTION : float
-    Thomson cross section [cm^2]
-PI : float
-    Ratio of a circle's circumference to its diameter
-"""
-
-# Constants in cgs units.
-AVOGADRO_NUMBER = 6.02214129e+23
-BOLTZMANN_CONSTANT = 1.3806488e-16
-CLASSICAL_ELECTRON_RADIUS = 2.8179402894e-13
-ELECTRONIC_CHARGE = 4.80320425e-10
-ELECTRON_VOLT = 1.602176565e-9
-ELECTRON_MASS = 9.10938188e-28
-FINE_STRUCTURE_CONSTANT = 7.2973525698e-3
-PLANCK_CONSTANT = 6.58211928e-19
-PROTON_MASS = 1.67261777e-24
-SPEED_OF_LIGHT = 299792458e+2
-THOMPSON_CROSS_SECTION = 6.652458734e-25
-PI = 3.14159265359
