+++ xdesign/algorithms.py
